+++ test/Common/RestClientTest.cs
@@ -66,7 +66,7 @@ namespace PasswordManagerAccess.Test.Common
         [Fact]
         public void Get_decodes_json()
         {
-            var response = Serve("{'Key': 'k', 'Value': 'v'}").Get<KeyValuePair<string, string>>(Url);
+            var response = Serve("""{"Key":"k","Value":"v"}""").Get<KeyValuePair<string, string>>(Url);
 
             Assert.True(response.IsSuccessful);
             Assert.Equal(new KeyValuePair<string, string>("k", "v"), response.Data);
@@ -156,6 +156,15 @@ namespace PasswordManagerAccess.Test.Common
         }
 
         [Fact]
+        public void PostJson_decodes_json()
+        {
+            var response = Serve("""{"Key":"k","Value":"v"}""").PostJson<KeyValuePair<string, string>>(Url, NoParameters);
+
+            Assert.True(response.IsSuccessful);
+            Assert.Equal(new KeyValuePair<string, string>("k", "v"), response.Data);
+        }
+
+        [Fact]
         public void PostRaw_sends_content_as_is()
         {
             var content = "blah-blah...";
@@ -202,7 +211,7 @@ namespace PasswordManagerAccess.Test.Common
         [Fact]
         public void Put_decodes_json()
         {
-            var response = Serve("{'Key': 'k', 'Value': 'v'}").Put<KeyValuePair<string, string>>(Url);
+            var response = Serve("""{"Key":"k","Value":"v"}""").Put<KeyValuePair<string, string>>(Url);
 
             Assert.True(response.IsSuccessful);
             Assert.Equal(new KeyValuePair<string, string>("k", "v"), response.Data);
@@ -214,6 +223,63 @@ namespace PasswordManagerAccess.Test.Common
             var response = Serve("", ResponseHeaders).Put(Url);
 
             Assert.Equal(ResponseHeaders, response.Headers);
+        }
+
+        //
+        // PUT JSON
+        //
+
+        [Fact]
+        public void PutJson_sends_json_headers()
+        {
+            InRequest(
+                rest => rest.PutJson(Url, NoParameters),
+                request => Assert.Equal(new[] { "application/json; charset=utf-8" }, request.Content.Headers.GetValues("Content-type"))
+            );
+        }
+
+        [Fact]
+        public void PutJson_encodes_json()
+        {
+            InRequest(
+                rest => rest.PutJson(Url, new Dictionary<string, object> { ["k"] = "v" }),
+                request => Assert.Equal("{\"k\":\"v\"}", request.Content.ReadAsStringAsync().Result)
+            );
+        }
+
+        [Fact]
+        public void PutJson_with_NoParameters_sends_empty_object()
+        {
+            InRequest(rest => rest.PutJson(Url, RestClient.NoParameters), request => Assert.Equal("{}", request.Content.ReadAsStringAsync().Result));
+        }
+
+        [Fact]
+        public void PutJson_with_JsonBlank_sends_blank()
+        {
+            InRequest(rest => rest.PutJson(Url, RestClient.JsonBlank), request => Assert.Equal("", request.Content.ReadAsStringAsync().Result));
+        }
+
+        [Fact]
+        public void PutJson_with_JsonNull_sends_null()
+        {
+            InRequest(rest => rest.PutJson(Url, RestClient.JsonNull), request => Assert.Equal("null", request.Content.ReadAsStringAsync().Result));
+        }
+
+        [Fact]
+        public void PutJson_returns_response_headers()
+        {
+            var response = Serve("", ResponseHeaders).PutJson(Url, NoParameters);
+
+            Assert.Equal(ResponseHeaders, response.Headers);
+        }
+
+        [Fact]
+        public void PutJson_decodes_json()
+        {
+            var response = Serve("""{"Key":"k","Value":"v"}""").PutJson<KeyValuePair<string, string>>(Url, NoParameters);
+
+            Assert.True(response.IsSuccessful);
+            Assert.Equal(new KeyValuePair<string, string>("k", "v"), response.Data);
         }
 
         //