+++ test/OnePassword/UtilTest.cs
@@ -2,7 +2,6 @@
 // Licensed under the terms of the MIT license. See LICENCE for details.
 
 using System;
-using Newtonsoft.Json;
 using PasswordManagerAccess.Common;
 using PasswordManagerAccess.Duo;
 using PasswordManagerAccess.OnePassword;
@@ -72,9 +71,8 @@ namespace PasswordManagerAccess.Test.OnePassword
         [Fact]
         public void DecryptAesKey_adds_decrypted_key_to_keychain()
         {
-            var encrypted = JsonConvert.DeserializeObject<R.Encrypted>(GetFixture("encrypted-aes-key"));
-            var keychain = new Keychain();
-            keychain.Add(new AesKey("mp", "44c38e8fedb84a1ab5ba74ed98dde931f6500ae39c1d9c85e20a7268ab2074f0".DecodeHex()));
+            var encrypted = ParseFixture<R.Encrypted>("encrypted-aes-key");
+            var keychain = new Keychain(new AesKey("mp", "44c38e8fedb84a1ab5ba74ed98dde931f6500ae39c1d9c85e20a7268ab2074f0".DecodeHex()));
 
             Util.DecryptAesKey(encrypted, keychain);
             var k = keychain.GetAes("szerdhg2ww2ahjo4ilz57x7cce").Key.ToHex();
@@ -85,9 +83,10 @@ namespace PasswordManagerAccess.Test.OnePassword
         [Fact]
         public void DecryptRsaKey_adds_decrypted_key_to_keychain()
         {
-            var encrypted = JsonConvert.DeserializeObject<R.Encrypted>(GetFixture("encrypted-rsa-key"));
-            var keychain = new Keychain();
-            keychain.Add(new AesKey("szerdhg2ww2ahjo4ilz57x7cce", "bba932f6032dc4dffaa9b8f03c9fd4b810127b89a49408db7b914a131690c091".DecodeHex()));
+            var encrypted = ParseFixture<R.Encrypted>("encrypted-rsa-key");
+            var keychain = new Keychain(
+                new AesKey("szerdhg2ww2ahjo4ilz57x7cce", "bba932f6032dc4dffaa9b8f03c9fd4b810127b89a49408db7b914a131690c091".DecodeHex())
+            );
 
             Util.DecryptRsaKey(encrypted, keychain);